enforce empty selection for @match
Summary:
This enforces that the match directive has an empty selection by directly converting to the MatchField IR type in the parser.

We can't easily do this conversion just in the transform because we also want to validate that normal non-scalar fields always have a selection. We thought about moving that validation later, but that doesn't work because other transforms automatically add `__typename` and `id` selection.

Reviewed By: jstejada

Differential Revision: D12872054

fbshipit-source-id: f8719b6345bedb96527eefea4ee4bce8c42fadfc