Allow compatible variable types with `@relay(mask: false)` (#2359)
Summary:
Changes the way variables/fragment arguments are validated for `relay(mask: false)`. Previously if the same variable was used in multiple unmasked fragments it had to have the exactly the same type. After this change, the most general type is inferred instead. So if `$id` is used in two locations, one expecting `ID` and one `ID!`, the argument will be inferred as the most general type `ID!` since that can flow into both locations. This also changes the location at which validation errors are thrown in order to find all variables that have validation errors within a fragment instead of throwing after the first validation error.

Fixes #2358

See https://github.com/facebook/relay/issues/2358 for details and an example.

I think this is safe to remove this. If the variables are actually not compatible it will be caught by the GraphQL validation step.
Pull Request resolved: https://github.com/facebook/relay/pull/2359

Reviewed By: jstejada

Differential Revision: D12890265

Pulled By: josephsavona

fbshipit-source-id: 487c7c3173b3ae1c05963cb3c39e9391d03bd196