Use new error reporting for ConnectionTransform
Summary: Uses the new RelayCompilerError utilities in ConnectionTransform (`create{User,Compiler}Error`). Because that transform uses IRTransformer, that means that errors from all uses of `connection` are now reported in aggregate, whereas before only the first invalid document was reported to the user. This should make it easier for developers to fix all errors in one iteration.

Reviewed By: kassens

Differential Revision: D13558577

fbshipit-source-id: f11f10829ce98cc688478c894c696193bd7c7dc9