Persisted queries (#2354)
Summary:
Enables a simple but straightforward mechanism to use persisted queries in open-source. Passing the new `--persist-output <path>` option now triggers the compiler to create an `id` for each operation (instead of leaving `id` null) and to save a mapping of ids to corresponding query text in the path specified by the developer. Developers can then save the id=>text mapping to a database or similar so that their GraphQL server can resolve query ids to text.

Note that the original PR saved a query mapping alongside each generated artifact and then aggregated those into a final query map, I've simplified this to only create a single aggregate query map at the end without any other intermediate artifacts. Also, instead of the `--persist` and optional `--persist-output` options, I've reduced this to just a single `--persist-output` argument so that the path to which the JSON mapping is saved is always user provided - this should help avoid confusion about where it ends up getting created and means that Relay compiler doesn't have to treat it as a generated file that may need to be cleaned up.

-- Original PR --

Hi guys, here's my attempt at implementing persisted queries for relay modern. Please review and let me know if it's ok.

Thanks!
Pull Request resolved: https://github.com/facebook/relay/pull/2354

Reviewed By: alunyov

Differential Revision: D13569254

Pulled By: josephsavona

fbshipit-source-id: cc10ce8b947016d1e3911a91230c44361ae98d9e