RefetchableFragmentTransform: use fragment spread in refetch query
Summary: Changes `RelayRefetchableFragmentTransform` to generate a refetch query that uses a fragment spread and passes all defined arguments, instead of inlining the selections of the original fragment. The idea is that we don't really need the fragment portion of the refetch operation since we can just use the original `refetchable` fragment to read data, so we might as well keep the refetch operation's fragment as minimal as possible.

Reviewed By: kassens

Differential Revision: D13577458

fbshipit-source-id: 13152d601dae2d838c186e48d119f7ba4293c49d