Rename OperationSelector to OperationDescriptor
Summary:
This is a proposal to rename `OperationSelector` to `OperationDescriptor`.

`OperationSelector` isn't exactly a selector in the same sense of a `ReaderSelector` or `NormalizationSelector`, instead it's essentially an aggregation of a bunch of data that you need to effectively execute that operation against a server with Relay.

I think the proposed name is slightly more accurate and helps avoid confusion about what a selector is.

Reviewed By: josephsavona

Differential Revision: D13723015

fbshipit-source-id: 470d91f9f6f8416d935aeb6ab933e156aa385172