Incremental 1/n: IR and Normalization kinds for @defer/@stream
Summary:
Adds new `Defer` and `Stream` IR and Normalization node kinds, along with updating all *existing* transforms/utilities to handle them correctly. Full tests for correctness are in the follow-up diff that transforms selections with `defer`/`stream` into these IR kinds, I'm breaking up the changes to aid review.

Notes on the design: Initially I considered creating a SplitOperation per label, and replacing defer/stream nodes with a "placeholder" (reference to the split operation e.g. by name/label). The idea for that approach is that there's not particular reason that the deferred/streamed selections themselves have to be present inline, since they'll be processed later anyway. There are some downsides though: the split selections still need to be associated with the node(s) that use them. Since defer/stream can appear in fragments, and be nested, that means traversing each query to see what defer/stream SplitOperations are reachable, and linking to them from the query. So it would end up being one transform to split the nodes, one to collect them back (two because we would also want to apply standard normalization transforms to the SplitOperations in between the split/collect stages). Keeping the selections inline means that it's just a simple transform, and every query can automatically reach all of its deferred/streamed selections. Another consideration is that removing a placeholder defer/stream means figuring out if the corresponding SplitOperation is still reachable or not - if the selections are inline, then the node can be trivially removed if appropriate.

More generally, ensuring that each IR kind has a distinct purpose helps to keep the system simple and adaptable. This is one of the reasons we have a distinct `Condition` kind - it means we can deal with conditional logic in one place, rather than have to check every single node's directives to look for conditionals. The same applies here - the notion of defer/stream is orthogonal to whether a node's selections are represented inline or not. We should perhaps add a concept of a non-inline selection (to exploit structural sharing at the IR level), but that's a separate project.

Next steps:
- Add a transform to create these nodes
- Start runtime support...

Reviewed By: kassens

Differential Revision: D13743344

fbshipit-source-id: b38e731d6a8d9a9d82650eafb33ad365f06714f2