Incremental 7/n: Normalizer accepts path prefix, returns match path
Summary: Changes ResponseNormalizer to accept an optional input `path` (on the options object). If present, this is used as the path prefix when normalizing selections (instead of starting with an empty path). This is necessary to support normalizing nested Match/Defer/Stream responses: processing a response requires knowing its absolute path in the server response, which means the path from a parent Match/Defer/Stream has to be fed back into Normalizer when processing a child response. The actual threading of paths will be implemented in a follow-up.

Reviewed By: kassens

Differential Revision: D13808278

fbshipit-source-id: d026023990fff4ae142bb813a7bda74549ce3aed