fix #2192 - dont recycle frozen objects (#2193)
Summary:
This is one possible fix for #2192, which simply checks if objects are frozen before attempting to modify them.

Do note that while this fixes my use case and any others that are serialized as simple JSON-compatible types, this won't fix cases where a custom scalar is [hydrated to a more complex type](https://github.com/facebook/relay/issues/91), such as a `Date` object. For this, you may want to consider adding a check to skip the recycle if the inputs aren't simple objects or arrays:

```es6
if (
  Object.getPrototypeOf(nextData) !== Object.getPrototypeOf(prevData) ||
  (
    Object.getPrototypeOf(nextData) !== Object.prototype &&
    Object.getPrototypeOf(nextData) !== Array.prototype
  )
) {
  return nextData;
}
```
But, since this may have more consequences, and I would really like my use case fixed ASAP 😜 I've omitted that "fix" from this PR. Please let me know if you'd like me to open a separate issue for that.
Pull Request resolved: https://github.com/facebook/relay/pull/2193

Reviewed By: josephsavona

Differential Revision: D6845724

Pulled By: jstejada

fbshipit-source-id: 2a96bb45b349334db51304bac393b7539ee00448