Incremental: QueryExecutor support for Defer
Summary:
Adds support for handling `defer` payloads in `QueryExecutor`. The executor now maintains a mapping of label => path => metadata about outstanding defer/stream payloads. When a deferred response arrives it the corresponding metadata is looked up and used to normalize it. The logic is basically the inverse of `match`: there, the data is ready ahead of time and it is the selector that has to built asynchronously. For `defer`, the selector is ready ahead of time and the data resolves asynchronously.

Next steps are to implement `stream`, test nested defer/stream/match, and then support handle incremental delivery of data within handle fields.

Reviewed By: jstejada

Differential Revision: D13865965

fbshipit-source-id: a85d07332657d065c6ac9ffe8ea81dbebddbdc59