Generalize @match representation, prepare to use @module independently
Summary:
This is the first of a few main steps towards generalizing the functionality of `match` to allow optionally using the lazy code/data loading functionality separately from the clients/server strategy-negotiation behavior. Today those two aspects of the functionality are coupled and `match` can only be used on fields that take a `supported` argument (to accomplish the strategy choice negotiation) and that return a union (to drive the laziness). There's no technical reason why they have to be coupled and we've found cases where it's beneficial to have just lazy loading. A simple example is a piece of content that has an optional attachment type (represented as a nullable linked field) that is almost always null. In this case, a developer may prefer to lazily load the code/data in the rare cases where the attachment is present to make the average (no attachment) case faster.

This diff changes the way `match` is represented in the IR/concrete asts, from a single `MatchField` node to a `LinkedField` containing `InlineFragments`. The `module` directive is currently still required on each selection, but in the future this directive will be handled independently from `match`. This keeps the concerns orthogonal and reduces a lot of duplication (note how all the traversals have much simpler handling of Match now, since they don't also have to redo field handling logic). This has a few ramifications that are reasonable but worth calling out:

* Normalization previously skipped processing a MatchField if the type didn't match any defined branches. Because `match` fields are just LinkedFields now, normalization will still normalize the field and simply find (when it reaches the Match) that none match and do nothing. This is arguably a more correct behavior, as it adds more information to the store and allows merging other selections under the same alias as the `match`.
* Reader similarly returned null for MatchField if the type didn't match, now it returns an empty object since the reader just sees it as a LinkedField where fields were missing. This will require some changes in `<MatchContainer>` when the diff is synced, e.g. requiring a fallback in case no branches match. Again, this is arguably better in the sense of encouraging writing resilient product code.
* The compiler now disallows having two `match` selections at the same point in the response, which matches existing behavior for fields with the same alias but different args.

The next step is to enable using `module` on fragment spreads to achieve only lazy loading.

Reviewed By: jstejada

Differential Revision: D13890352

fbshipit-source-id: 655fcb7ddb7b9166c6c9e3de0fda71d4a780a13f