RM Owners: Include fetchVariables in fragmentVariables when refetching
Summary:
Currently, when paginating or refetching inside our Refetch and Pagination Containers there are 2 sets of variables at play:

- The `fetchVariables`, which represent the operation variables that will be used to fetch a refetch/pagination query from the server,
- The `fragmentVariables` which are essentially a user-provided set of variables spread over the original root operation variables. When a refetch occurs, these "fragmentVariables" are used to 1) read the local fragment, and 2) are set in Context and propagated down to descendants as the new "rootVariables", meaning that descendant containers will use them as input to read their own fragments.

Note that:
- Usually when creating a selector and computing the fragment variables, we take into account the root operation variables + the variables available in the fragment ref (i.e. which include variables passed as arguments). However, the `fragmentVariables` computed for refetch are user provided and ignore the fragment refs that would've been obtained from the refetch query response (also ignoring any arguments), so they're not guaranteed to be correct.
- This fragmentVariables` and `fetchVariables` are often different sets of variables, with some or no overlap, which means that descendants can't always access the variables used to fetch the operation.

Confusingly, Pagination and Refetch Container also have slightly different logic for computing these variables:
- In the pagination container, the `fragmentVariables` are computed from the original fragment refs, and then used as input to a user-provided function to construct the `fetchVariables`. The fragment variables are set in Context, but then later modified by another user-provided function before using them to read the local fragment.
- In the Refetch Container, both `fetchVariables` and `fragmentVariables` are just directly provided by the user, and if `fragmentVariables` aren't provided they just default to the `fetchVariables`.

Given this multitude of behaviors when managing variables it's hard to tell exactly which variables should be used when, and know what the **"correct"** behavior is. I believe that developers using or constructing these variables also don't know exactly what the correct behavior is, and the intended behavior to separate "fetching" from "reading" is misused or not really used.

That being said, the purpose of this commit **IS NOT** to address the fundamental issues with the existing refetch/pagination containers. *Instead*, the purpose of this commit is to enable us to preserve the existing broken behavior within a fragment ownership-based model. Essentially, this commit is a stepping stone that will allow RelayModern containers to both set and **consume** variables determined by fragment owners instead of propagating variables with React Context.

As to *why* we want to enable fragment ownership in RelayModern containers:

- It removes our reliance on React Context to propagate variables and allows Relay Modern containers to be compatible with other containers that are based on fragment ownership and not Context (https://fb.quip.com/2stcADv99pad).
- It will allow us to address the issues around variable management and move towards a more correct model.

This commit makes the variables passed down in Context a union of both `fragmentVariables` and `fetchVariables`, where the values of the `fragmentVariables` are preferred over the `fetchVariables`.

**Why do we need to do this?**

Using fragment ownership, refetching or paginating conceptually means re-parenting a fragment under a new owning operation; that is, the variables from the owning operation + the variables from the new fragment refs would determine the variables used to read a fragment, instead of using variables read from React Context. This means that in order to preserve existing behavior, the variables propagated via a fragment owner need to :
- Contain what are currently known as the `fragmentVariables.
- Also actually contain the `fetchVariables` used to fetch the operation, since this information will also be used to access information about the operation, e.g. if it is currently in flight.

This change should be safe since it's purely additive, meaning that the `fragmentVariables` will only possibly contain extra variables, and any of the values currently used in the `fragmentVariables` will remain unchanged at runtime.

Verifying that this new behavior is still correct and doesn't affect product code will give us the verification needed to start actually setting and consuming fragment owners with these variables (i.e. the union of `fragmentVariables` and `fetchVariables`), and replace them for the variables set in context.

**Additionally, the change is gated under a feature flag to be able to easily revert it if anything breaks.**

- Start consuming fragment owners within RelayModern (next diff in this stack)
- Once containers are propagating variables via fragment owners instead of context, we can work on making this behavior actually correct: Even if we can't literally extract the fragment refs from a refetch response, it should still produce the same result: that the fragments have as owner a query with just fetchVariables, and have the appropriate fragment variables as if they had been extracted from the query result (e.g. considering possible arguments).

Reviewed By: josephsavona

Differential Revision: D13973921

fbshipit-source-id: 61b254714c170f751b361083a012bde2e703e100