RM Owners: Relax constraint for fragment owner being non-null in RelayModern
Summary:
Currently, whenever we call getFragmentOwners we expect the fragment ref to have a non-null owner attached to it or we throw an error. Essentially, we expect that the object that is passed as the fragment ref is actually a fragment ref. Similarly, when calling getSelectorsFromObject, if we passed a map of fragment owners, we expected them to be defined, or otherwise we'd throw an error.

However, this isn't compatible with how RelayModern containers currently work, since it is possible to pass non-ref objects in place of fragment refs to be used as mock data for rendering in tests. That is, if we pass an object that isn't a fragment ref where a fragment ref is expected, we will simply pass it down as is, meaning you can put mock data in it.

In order to continue supporting this, the fragment ownership helpers can't throw an error if the object isn't actually a fragment ref, and just return null so the object is just passed through.

Reviewed By: josephsavona

Differential Revision: D14012043

fbshipit-source-id: ce07975434b6d83d11a727499da1b405590d4aae