2/2: Expose getSelector and getVariablesFromFragment helpers
Summary:
This commit exposes new helper functions `getSelector` and `getVariablesFromFragment`, which know how tu return either singular or plural results and have proper validation.

These are handy for clients that want to reuse the validation from `getSelectorsFromObject` and `getVariablesFromObject` when dealing with a single fragment, without having to unnecessarily wrap it in an object to call the `*FromObject` version.

Reviewed By: josephsavona

Differential Revision: D14112823

fbshipit-source-id: 1be0288786dbddf083035fc98c2df0575ed01438