Pass the IR definition to the generated file writer
Summary: Adds a `definition` parameter to the `FormatModule` type, allowing generated file writers access to the IR definition of the generated node in addition to the hash, code text, type text, etc. This is useful for generated additional metadata - the concrete use-case is to write out the possible set of dependencies that may be loaded via `match` / `module`.

Reviewed By: kassens

Differential Revision: D14157564

fbshipit-source-id: 5cd21ef01823b75c76652501ced7d446fdee8945